refactor: extract handlers into handler.go (no API change)
Engineering review item #4: bolt.go was a 2002-line single file mixing
encoders, handlers, validation, the event pool, and the public Logger.
Move the handler types and their JSON-streaming helpers into a focused
handler.go (352 lines) without changing any exported API or behaviour.

What moved out of bolt.go:
  * JSONHandler, ConsoleHandler, multiHandler types
  * NewJSONHandler, NewConsoleHandler, MultiHandler constructors
  * Their Write methods (mutex-serialised after #48)
  * findJSONFieldStart, extractStringValue, extractNonStringValue,
    extractJSONField, extractJSONKey, extractJSONValue
  * skipWhitespace, skipCommaIfPresent
  * writeKeyValue, writeFieldsStreaming, isReservedField
  * getColorForLevel
  * Drops the now-unused `bytes` import from bolt.go

What stayed in bolt.go:
  * Event type, Logger type, Hook + ErrorHandler interfaces
  * Level constants, level aliases
  * appendInt / appendUint / appendBool / appendJSONString /
    appendIP / appendRFC3339 / appendFloat64 / etc.
  * eventPool, default logger, environment-variable parsing
  * NewLevelWriter

handler.go is the natural first split because handlers are a
self-contained concern: they don't poke into Event internals beyond
e.buf, and they're the entry point downstream consumers most often
need to read when implementing custom handlers.

Pure code-motion. No symbols added or removed; no public API change.

Verification:
  * `go test -short -count=1 ./...` — passes
  * `go test -race -short -count=1 .` — passes
  * `golangci-lint run` — 0 issues (one staticcheck QF1001
    De-Morgan's-law cleanup applied while moving the affected loop)
  * `BenchmarkZeroAllocation` — 56 ns/op, 0 allocs/op (within budget)
  * bolt.go: 2002 → 1617 lines; handler.go: new 352-line file
